First commit towards Sigar implemenation
Initial commit for Sigar implementation. Added SigarProcessManager and some tests.
I have added some TODO's with some questions etc that would be nice to get feedback on

Remember to read the TODO and README. To run the tests you need to add the depends folder in the java.library.path or LD_LIBRARY_PATH enviroment